backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-jose==3.3.0
python-dotenv==1.0.1
openai==1.43.0
chromadb==0.5.5
httpx==0.27.2 
httpcore==1.0.5
better-profanity==0.7.0
oracledb==2.2.1
SpeechRecognition==3.10.4
pyttsx3==2.90